refactor console
To improve the code, the commands are now distributed over several
separate files. This also adds a "database" command to
backup/restore a database/dump or execute sql queries without
messing around with credentials.

Note, we trust the machine which
executes the infomark binary and also stores the config. Hence, this
does not introduce a security issue. It is just a neat wrapper.