bypass login for local deployment
Creating an optimized elm production build via `elm-app build` takes some time. To use `elm-app start` without any login procedure we add a `debug_user_id` flag with default value "0". If this value is not equal "0", the id is used as the loginID granted the any request full root access to the system. The intended usage is really just for local development.